Add auth for gigs

+++ gigServer/src/controllers/gigs.js
@@ -1,5 +1,6 @@
 const GigsModel = require("../models/Gigs");
 const ProviderAuthModel = require("../models/ProviderAuth");
+const jwt = require("jsonwebtoken");
 
 const getAllGigs = async (req, res) => {
   try {
@@ -24,27 +25,34 @@ const getGigById = async (req, res) => {
 // provider add gig
 const addGigForProvider = async (req, res) => {
   try {
-    const provider = await ProviderAuthModel.findById(req.params.providerId);
+    // Double check if profile exist and matches
+    if (req.decoded.id != req.params.providerId) {
+      return res
+        .status(400)
+        .json({ status: "error", msg: "error finding profile to patch" });
+    } else {
+      const provider = await ProviderAuthModel.findById(req.params.providerId);
 
-    // create a new GigsModel instance with the provided data
-    const newGig = new GigsModel({
-      title: req.body.title,
-      author: provider._id,
-      dateTimeStart: req.body.dateTimeStart,
-      pic: req.body.pic,
-      address: req.body.address,
-      link: req.body.link,
-      description: req.body.description,
-    });
+      // create a new GigsModel instance with the provided data
+      const newGig = new GigsModel({
+        title: req.body.title,
+        author: provider._id,
+        dateTimeStart: req.body.dateTimeStart,
+        pic: req.body.pic,
+        address: req.body.address,
+        link: req.body.link,
+        description: req.body.description,
+      });
 
-    // save the data to the GigsModel
-    await newGig.save();
-    // add the newly created gig OID to the hostGigsList array in providerAuth collection
-    provider.hostGigsList.push(newGig._id);
-    // save the collection
-    await provider.save();
+      // save the data to the GigsModel
+      await newGig.save();
+      // add the newly created gig OID to the hostGigsList array in providerAuth collection
+      provider.hostGigsList.push(newGig._id);
+      // save the collection
+      await provider.save();
 
-    res.json({ status: "ok", msg: "gig created" });
+      res.json({ status: "ok", msg: "gig created" });
+    }
   } catch (error) {
     console.error(error.message);
     res.status(400).json({ status: "error", msg: "add gig unsuccessful" });
@@ -53,16 +61,25 @@ const addGigForProvider = async (req, res) => {
 
 const deleteGigForProvider = async (req, res) => {
   try {
-    const provider = await ProviderAuthModel.findById(req.params.providerId);
+    const gig = await GigsModel.findById(req.params.id);
 
-    await GigsModel.findByIdAndDelete(req.params.id);
+    // Double check if profile exist and matches
+    if (req.decoded.id != gig.author) {
+      return res
+        .status(400)
+        .json({ status: "error", msg: "error finding profile to patch" });
+    } else {
+      const provider = await ProviderAuthModel.findById(req.decoded.id);
 
-    // delete gig OID from hostGigsList array in providerAuth collection
-    const indexToDelete = provider.hostGigsList.indexOf(req.params.id);
-    provider.hostGigsList.splice(indexToDelete, 1);
-    await provider.save();
+      await GigsModel.findByIdAndDelete(req.params.id);
 
-    res.json({ status: "ok", msg: "gig deleted" });
+      // delete gig OID from hostGigsList array in providerAuth collection
+      const indexToDelete = provider.hostGigsList.indexOf(req.params.id);
+      provider.hostGigsList.splice(indexToDelete, 1);
+      await provider.save();
+
+      res.json({ status: "ok", msg: "gig deleted" });
+    }
   } catch (error) {
     console.error(error.message);
     res.status(400).json({ status: "error", msg: "delete gig unsuccessful" });
@@ -71,17 +88,27 @@ const deleteGigForProvider = async (req, res) => {
 
 const updateGigForProvider = async (req, res) => {
   try {
-    const updateGig = {};
-    if ("title" in req.body) updateGig.title = req.body.title;
-    if ("dateTimeStart" in req.body)
-      updateGig.dateTimeStart = req.body.dateTimeStart;
-    if ("pic" in req.body) updateGig.pic = req.body.pic;
-    if ("address" in req.body) updateGig.address = req.body.address;
-    if ("link" in req.body) updateGig.link = req.body.link;
-    if ("description" in req.body) updateGig.description = req.body.description;
+    const gig = await GigsModel.findById(req.params.id);
 
-    await GigsModel.findByIdAndUpdate(req.params.id, updateGig);
-    res.json({ status: "ok", msg: "gig updated" });
+    // Double check if profile exist and matches
+    if (req.decoded.id != gig.author) {
+      return res
+        .status(400)
+        .json({ status: "error", msg: "error finding profile to patch" });
+    } else {
+      const updateGig = {};
+      if ("title" in req.body) updateGig.title = req.body.title;
+      if ("dateTimeStart" in req.body)
+        updateGig.dateTimeStart = req.body.dateTimeStart;
+      if ("pic" in req.body) updateGig.pic = req.body.pic;
+      if ("address" in req.body) updateGig.address = req.body.address;
+      if ("link" in req.body) updateGig.link = req.body.link;
+      if ("description" in req.body)
+        updateGig.description = req.body.description;
+
+      await GigsModel.findByIdAndUpdate(req.params.id, updateGig);
+      res.json({ status: "ok", msg: "gig updated" });
+    }
   } catch (error) {
     console.error(error.message);
     res.status(400).json({ status: "error", msg: "update gig unsuccessful" });
